More delegate methods: (1) When search results return exactly one matching item (2) When user taps on search result in list (3) When user taps on map and geocode returns a result (4) User selects an annotation on the map that was added from a search result.
Map improvement: when user taps on search table entry, hide the table and zoom the map to that location

UX improvements: (1) added a public accessible ratio to control how large the expanded stack can be relative to the height of the view. (2) Don’t popup alert dialog when user clicks on the map and geocode search comes back - instead let delegate decide what to do. (3) Removed code that causes an undesirable behavior: when user scrolls in the table view and hits the top, the table slides out of view.